moved code to run native games to spel-katalog-run
